+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 ### Added
+- Extract query now supports extractFormat
 
 ### Fixed
 
 ### Changed
+- Exception message for invalid/unavailable file in Extract query now contains filename
 
 ### Removed